Refactor App & Canvas, use /public dependencies instead of CDNs.
This fixes the unnecessary usage of React Router just to inject the painting ID
by simply passing it as a prop to the #root div that the Canvas component mounts
to. Because the Router is redundant, the App component is also redundant and can
be replaced directly by the router.
Also, the CDN sources for Bootstrap and its dependencies have been replaced by
the pre-compiled versions in the /public dir.