Add test for setting painting title, fix bug in PaintingUpdateEvent
The `action` field was improperly set to be a `string`, when in some cases the
`action` field should be null - one such case is when a User sends a
PaintingUpdateRequest to change the title of a Painting. In this case, the
`action` field is missing and only the `title` is set.
To fix this, PHP 8 union types would be ideal because the type of `action` could
be `string|null`. In the meantime, however, I'll just elide the type annotation
for the `action` field.